+++ packages/web/server/queries/base-utils.ts
@@ -1,12 +1,9 @@
 import { Chain } from "@osmosis-labs/types";
-import { apiClient, getChainRestUrl, isNil } from "@osmosis-labs/utils";
-import { createClient, VercelKV } from "@vercel/kv";
-import { Cache, CacheEntry, totalTtl } from "cachified";
+import { apiClient, getChainRestUrl } from "@osmosis-labs/utils";
 import DataLoader from "dataloader";
 
 import { ChainList } from "~/config/generated/chain-list";
 import { runIfFn } from "~/utils/function";
-import { superjson } from "~/utils/superjson";
 
 /** Batching DataLoader with Vercel Edge compatible batching scheduler.
  *
@@ -27,46 +24,6 @@ export class EdgeDataLoader<K, V, C = K> extends DataLoader<K, V, C> {
       batchScheduleFn: (cb) => setTimeout(cb, 0),
       ...args[1],
     } as ConstructorParameters<typeof DataLoader<K, V, C>>[1]);
-  }
-}
-
-// inspired by: https://github.com/mannyv123/cachified-redis-adapter/blob/main/src/index.ts
-/** `cachified`-compatible implementation of a cache living on a remote resource. */
-export class RemoteCache implements Cache {
-  protected kvStore: VercelKV = createClient({
-    url: process.env.TWITTER_KV_STORE_REST_API_URL!,
-    token: process.env.TWITTER_KV_STORE_REST_API_TOKEN!,
-  });
-
-  name = "RemoteCache";
-
-  async get(key: string) {
-    const value = await this.kvStore.get(key);
-    if (isNil(value) || typeof value !== "string") {
-      return null;
-    }
-    // Note that parse can potentially throw an error here and the expectation is that the user of the adapter catches it
-    return superjson.parse(value) as CacheEntry<unknown>;
-  }
-
-  async set(key: string, value: CacheEntry<unknown>) {
-    const ttl = totalTtl(value?.metadata);
-    const createdTime = value?.metadata?.createdTime;
-
-    await this.kvStore.set(
-      key,
-      superjson.stringify(value),
-      ttl > 0 && ttl < Infinity && typeof createdTime === "number"
-        ? {
-            // convert the exat to seconds by dividing by 1000
-            exat: Math.ceil((ttl + createdTime) / 1000),
-          }
-        : undefined
-    );
-  }
-
-  async delete(key: string) {
-    await this.kvStore.del(key);
   }
 }
